refactor: mv src/gui/src/AddClientDialog => src/gui/src/dialog/AddClientDialog
