+++ src/gui/src/dialogs/AddClientDialog.cpp
