+++ src/gui/src/dialogs/AddClientDialog.h
