+++ src/gui/src/dialogs/AddClientDialog.ui
